LGA Taxi Demand & Supplly Simulation
New York City’s LaGuardia Airport (LGA) is notorious for its mismatch in the demand and supply of taxis, with passengers often waiting over and hour, or taxis waiting of several hours.  Via, a ride-hailing app (similar to Uber) headquartered in New York City, has recently started servicing rides to and from LaGuardia airport and wants to increase its quality of service (minimizing wait times and unserved requests) while maximizing profits (minimizing number of cars sent and driver working hours). 
This simulation aims to model the potential dynamics of queueing of passengers and taxis in LGA airport and transporting into (and from) Manhattan.  It aims to model with some resemblance to reality the following processes:
1.	Patterns of passenger exit times from LGA airport and ordering service requests
2.	Patterns of drop-offs occurring naturally at LGA airport
3.	The company’s strategy for forcibly sending cars to wait at LGA queue or away from it to optimize parameters.